+++ ncollide_entities/shape/composite_shape.rs
@@ -12,41 +12,41 @@ use inspection::Repr;
 /// a concave shape.
 pub trait CompositeShape<N, P, V, M> {
     /// Applies a function to each sub-shape of this concave shape.
-    fn map_part_at(&self, uint, &mut FnMut(&M, &Repr<N, P, V, M>));
+    fn map_part_at(&self, usize, &mut FnMut(&M, &Repr<N, P, V, M>));
     /// Applies a transformation matrix and a function to each sub-shape of this concave
     /// shape.
-    fn map_transformed_part_at(&self, m: &M, uint, &mut FnMut(&M, &Repr<N, P, V, M>));
+    fn map_transformed_part_at(&self, m: &M, usize, &mut FnMut(&M, &Repr<N, P, V, M>));
 
     // FIXME: the following two methods really are not generic enough.
     /// Gets the AABB of the shape identified by the index `i`.
-    fn aabb_at(&self, i: uint) -> &AABB<P>;
+    fn aabb_at(&self, i: usize) -> &AABB<P>;
     /// Gets the acceleration structure of the concave shape.
-    fn bvt(&self) -> &BVT<uint, AABB<P>>;
+    fn bvt(&self) -> &BVT<usize, AABB<P>>;
 }
 
 impl<N, P, V, M> CompositeShape<N, P, V, M> for Compound<N, P, V, M>
     where M: Copy + Mul<M, Output = M> {
     #[inline(always)]
-    fn map_part_at(&self, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_part_at(&self, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let &(ref m, ref g) = &self.shapes()[i];
 
         f(m, &***g)
     }
 
     #[inline(always)]
-    fn map_transformed_part_at(&self, m: &M, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_transformed_part_at(&self, m: &M, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let elt = &self.shapes()[i];
 
         f(&(*m * elt.0), &**elt.1)
     }
 
     #[inline]
-    fn aabb_at(&self, i: uint) -> &AABB<P> {
+    fn aabb_at(&self, i: usize) -> &AABB<P> {
         &self.bounding_volumes()[i]
     }
 
     #[inline]
-    fn bvt(&self) -> &BVT<uint, AABB<P>> {
+    fn bvt(&self) -> &BVT<usize, AABB<P>> {
         self.bvt()
     }
 }
@@ -57,26 +57,26 @@ impl<N, P, V, M> CompositeShape<N, P, V, M> for TriMesh<N, P, V>
           V: Vect<N>,
           M: Isometry<N, P, V> {
     #[inline(always)]
-    fn map_part_at(&self, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_part_at(&self, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let one: M = na::one();
 
         self.map_transformed_part_at(&one, i, f)
     }
 
     #[inline(always)]
-    fn map_transformed_part_at(&self, m: &M, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_transformed_part_at(&self, m: &M, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let element = self.triangle_at(i);
 
         f(m, &element)
     }
 
     #[inline]
-    fn aabb_at(&self, i: uint) -> &AABB<P> {
+    fn aabb_at(&self, i: usize) -> &AABB<P> {
         &self.bounding_volumes()[i]
     }
 
     #[inline]
-    fn bvt(&self) -> &BVT<uint, AABB<P>> {
+    fn bvt(&self) -> &BVT<usize, AABB<P>> {
         self.bvt()
     }
 }
@@ -87,26 +87,26 @@ impl<N, P, V, M> CompositeShape<N, P, V, M> for Polyline<N, P, V>
           V: Vect<N>,
           M: Isometry<N, P, V> {
     #[inline(always)]
-    fn map_part_at(&self, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_part_at(&self, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let one: M = na::one();
 
         self.map_transformed_part_at(&one, i, f)
     }
 
     #[inline(always)]
-    fn map_transformed_part_at(&self, m: &M, i: uint, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
+    fn map_transformed_part_at(&self, m: &M, i: usize, f: &mut FnMut(&M, &Repr<N, P, V, M>)) {
         let element = self.segment_at(i);
 
         f(m, &element)
     }
 
     #[inline]
-    fn aabb_at(&self, i: uint) -> &AABB<P> {
+    fn aabb_at(&self, i: usize) -> &AABB<P> {
         &self.bounding_volumes()[i]
     }
 
     #[inline]
-    fn bvt(&self) -> &BVT<uint, AABB<P>> {
+    fn bvt(&self) -> &BVT<usize, AABB<P>> {
         self.bvt()
     }
 }
