+++ ncollide_entities/shape/polyline.rs
@@ -10,7 +10,7 @@ use math::{Scalar, Point, Vect};
 
 /// Shape commonly known as a 2d line strip or a 3d segment mesh.
 pub struct Polyline<N, P, V> {
-    mesh: BaseMesh<N, P, V, Pnt2<uint>, Segment<P>>
+    mesh: BaseMesh<N, P, V, Pnt2<usize>, Segment<P>>
 }
 
 impl<N, P, V> Clone for Polyline<N, P, V>
@@ -28,7 +28,7 @@ impl<N, P, V> Polyline<N, P, V>
           V: Translate<P> + Vect<N> {
     /// Builds a new mesh.
     pub fn new(vertices: Arc<Vec<P>>,
-               indices:  Arc<Vec<Pnt2<uint>>>,
+               indices:  Arc<Vec<Pnt2<usize>>>,
                uvs:      Option<Arc<Vec<Pnt2<N>>>>,
                normals:  Option<Arc<Vec<V>>>) // a loosening margin for the BVT.
                -> Polyline<N, P, V> {
@@ -41,7 +41,7 @@ impl<N, P, V> Polyline<N, P, V>
 impl<N, P, V> Polyline<N, P, V> {
     /// The base representation of this mesh.
     #[inline]
-    pub fn base_mesh(&self) -> &BaseMesh<N, P, V, Pnt2<uint>, Segment<P>> {
+    pub fn base_mesh(&self) -> &BaseMesh<N, P, V, Pnt2<usize>, Segment<P>> {
         &self.mesh
     }
 
@@ -59,7 +59,7 @@ impl<N, P, V> Polyline<N, P, V> {
 
     /// The indices of this mesh.
     #[inline]
-    pub fn indices(&self) -> &Arc<Vec<Pnt2<uint>>> {
+    pub fn indices(&self) -> &Arc<Vec<Pnt2<usize>>> {
         unsafe { mem::transmute(self.mesh.indices()) }
     }
 
@@ -77,7 +77,7 @@ impl<N, P, V> Polyline<N, P, V> {
 
     /// The acceleration structure used for efficient collision detection and ray casting.
     #[inline]
-    pub fn bvt(&self) -> &BVT<uint, AABB<P>> {
+    pub fn bvt(&self) -> &BVT<usize, AABB<P>> {
         self.mesh.bvt()
     }
 }
@@ -85,7 +85,7 @@ impl<N, P, V> Polyline<N, P, V> {
 impl<N, P: Send + Sync + Copy + Dim, V> Polyline<N, P, V> {
     /// Gets the i-th mesh element.
     #[inline]
-    pub fn segment_at(&self, i: uint) -> Segment<P> {
+    pub fn segment_at(&self, i: usize) -> Segment<P> {
         self.mesh.element_at(i)
     }
 }
